#698 - refactor to be minimal change at import only, other than tests, and attempt to remove that from coverage
