+++ README.md
@@ -39,7 +39,8 @@ The compatible versions of [Spark](http://spark.apache.org/) and
 | 0.11.1    | 3.2.0 / 3.1.2 / 3.0.1         | 2.x      | 2.x         | 2.12 / 2.13
 | 0.12.0    | 3.2.1 / 3.1.3 / 3.0.3         | 2.x      | 3.x         | 2.12 / 2.13
 | 0.13.0    | 3.3.0 / 3.2.2 / 3.1.3         | 2.x      | 3.x         | 2.12 / 2.13
-| 0.14.0    | 3.3.0 / 3.2.2 / 3.1.3 / 3.4.0 | 2.x      | 3.x         | 2.12 / 2.13
+| 0.14.0    | 3.3.0 / 3.2.2 / 3.1.3         | 2.x      | 3.x         | 2.12 / 2.13
+| 0.14.1    | 3.3.0 / 3.2.2 / 3.1.3 / 3.4.0 | 2.x      | 3.x         | 2.12 / 2.13
 
 _\* 0.11.0 has broken Spark 3.1.2 and 3.0.1 artifacts published._
